Introduce a dependency on Abseil
The main motivation is enabling the benchmark library to play well with
hosts that use Abseil, especially for command line flag functionality.
This also opens up the possibility of reusing functionality such as
string utils, concurrency, or logging.

This was previously discussed in issue #910. This patch:

- adds the dependency to Abseil. The library will either be downloaded
(at a fixed, specific commit) or the user may specify their own location
via the ABSEIL_PATH cmake flag

- replaces flag functionality with Abseil's.